requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
boto3>=1.26.0
pypdf>=3.0.0
python-docx>=1.2.0


